% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{data_overDue}
\alias{data_overDue}
\title{排产数据查询}
\usage{
data_overDue(FToken = "7434C334-6E04-43C7-B103-634963BF78EF")
}
\arguments{
\item{FToken}{口令}
}
\value{
返回值
}
\description{
排产数据查询
}
\examples{
data_query()
}
