% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{data_query}
\alias{data_query}
\title{排产数据查询}
\usage{
data_query(FToken = "7434C334-6E04-43C7-B103-634963BF78EF", status = "open")
}
\arguments{
\item{FToken}{口令}

\item{status}{状态}
}
\value{
返回值
}
\description{
排产数据查询
}
\examples{
data_query()
}
