% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{data_upload}
\alias{data_upload}
\title{上传数据}
\usage{
data_upload(
  FToken = "7434C334-6E04-43C7-B103-634963BF78EF",
  table_name = "rds_scheduler",
 
    file_name = "data-raw/棱星数据项目管理套表_W21_v1.3_胡立磊_20230520.xlsx",
  sheet = "data",
  type = "WBS"
)
}
\arguments{
\item{FToken}{口令}

\item{table_name}{表名}

\item{file_name}{文件名}

\item{sheet}{页签名}

\item{type}{类型}
}
\value{
返回值
}
\description{
上传数据
}
\examples{
data_upload()
}
