% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{dataDemo_check}
\alias{dataDemo_check}
\title{查询排产数据}
\usage{
dataDemo_check(FToken = "7434C334-6E04-43C7-B103-634963BF78EF", type = "WBS")
}
\arguments{
\item{FToken}{增加口令}

\item{type}{增加业务类型}
}
\value{
返回值
}
\description{
查询排产数据
}
\examples{
dataDemo_check()
}
