% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.R
\name{projectFormatter}
\alias{projectFormatter}
\title{项目状态格式化格式}
\usage{
projectFormatter(id, task, status)
}
\arguments{
\item{id}{项目内码}

\item{task}{项目任务}

\item{status}{状态}
}
\value{

}
\description{
项目状态格式化格式
}
\examples{
返回值
projectFormatter()
}
