% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_close}
\alias{task_close}
\title{任务关闭}
\usage{
task_close(
  FToken = "7434C334-6E04-43C7-B103-634963BF78EF",
  FId = "A0027",
  FStart,
  FEnd,
  FNote
)
}
\arguments{
\item{FToken}{口令}

\item{FId}{任务内码}

\item{FStart}{实际开始时间}

\item{FEnd}{实际完成时间}

\item{FNote}{完成情况}
}
\value{
返回值
}
\description{
任务关闭
}
\examples{
task_close()
}
