% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{taskId_defautValue}
\alias{taskId_defautValue}
\title{获取默认值}
\usage{
taskId_defautValue(
  FToken = "7434C334-6E04-43C7-B103-634963BF78EF",
  type = "WBS"
)
}
\arguments{
\item{FToken}{口令}

\item{type}{类型}
}
\value{
返回值
}
\description{
获取默认值
}
\examples{
taskId_defautValue()
}
