% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{taskId_NextValue}
\alias{taskId_NextValue}
\title{获取单据代码的下一个值}
\usage{
taskId_NextValue(FToken = "7434C334-6E04-43C7-B103-634963BF78EF", type = "QA")
}
\arguments{
\item{FToken}{口令}

\item{type}{类型}
}
\value{
返回值
}
\description{
获取单据代码的下一个值
}
\examples{
taskId_NextValue()
}
