% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{taskId_valueLen}
\alias{taskId_valueLen}
\title{获取前缀}
\usage{
taskId_valueLen(FToken = "7434C334-6E04-43C7-B103-634963BF78EF", type = "WBS")
}
\arguments{
\item{FToken}{口令}

\item{type}{类型}
}
\value{
返回值
}
\description{
获取前缀
}
\examples{
taskId_valueLen()
}
